refactor: t2i_diffusion_benchmark.metrics module
