+++ t2i_diffusion_benchmark/metrics/prompt_alignment/__init__.py
@@ -1,0 +1,5 @@
+from .blip_scorer import BLIPScorer
+from .clip_iqa_scorer import CLIPImageQualityScorer
+from .clip_scorer import CLIPScorer
+
+__all__ = ["BLIPScorer", "CLIPScorer", "CLIPImageQualityScorer"]
